Offload the LoadOrdering to caller
While a full end-user experience would order the desired mods according to their order in plugin.txt, we shouldn't do that within the API that is concerned with the minimal job of merging.  Doing so would never allow anyone to merge according to their own specific ordering.

Just let the caller give the mods and the order they want.  If they want to conform to plugin.txt (which a user-facing app would), they can just use the LoadOrder tooling to do so before passing in